+++ app/classes/exhibitPresenter.jsx
@@ -3,11 +3,15 @@ import { Link, useLoaderData } from "@remix-run/react"
 import { useEffect } from 'react'
 import { renderAuthorBubble, renderPageLink, renderPageLinks, renderSidebar, renderSidebarSection, renderPageTitleBar, renderFootnoteSection, renderFootnotesInBody } from "./pageHelpers"
 import { renderBlocks, renderBlock, textContent, headingContent } from "./contentHelpers"
+import { SIDEBAR_TYPES } from '../data/sidebarTypes'
 
 export function renderExhibit(exhibit){
   let showAuthors = exhibit.authors && exhibit.authors.length > 0 && exhibit.authors[0].name
   let showFootnotes = exhibit.footnotes && exhibit.footnotes.length > 0
-  let sidebar = renderSidebar("In This Exhibit", exhibit.body.filter( (block) => block.type == "heading"), showAuthors, showFootnotes)  
+
+  // previously only targeted heading for exhibit sidebars, but changed to match collections
+  // let sidebar = renderSidebar("In This Exhibit", exhibit.body.filter( (block) => block.type == "heading"), showAuthors, showFootnotes)  
+  let sidebar = renderSidebar("In This Exhibit", exhibit.body.filter( (block) => SIDEBAR_TYPES.includes(block.type)), showAuthors, showFootnotes)  
 
   let titleBar
   if(exhibit.title){
