+++ app/data/aapbBlockTypes.jsx
@@ -1,0 +1,8 @@
+export const AAPB_BLOCK_TYPES = [
+  'interviews',
+  'archival_footage',
+  'photographs',
+  'original_footage',
+  'programs',
+  'related_content',
+]
