+++ app/data/sidebarTypes.jsx
@@ -1,0 +1,10 @@
+export const SIDEBAR_TYPES = [
+  'heading',
+  'interviews',
+  'archival_footage',
+  'photographs',
+  'original_footage',
+  'programs',
+  'related_content',
+  'credits',
+]
