+++ app/routes/preview.tsx
@@ -4,9 +4,27 @@ import { getPreview } from '../preview'
 import { renderCollection } from '../classes/collectionPresenter'
 import { renderExhibit } from '../classes/exhibitPresenter'
 import type { SitemapFunction } from 'remix-sitemap'
+import { getPageBySlug } from '../fetch'
 
 export const loader = async ({ request }: LoaderFunctionArgs) => {
   var params = new URLSearchParams(request.url.replace(/.*\?/, ''))
+  // if(params.has("slug") && params.has("token") && params.has("content_type")){
+
+  //   var slug = params.get("slug")
+  //   var token = params.get("token")
+  //   var content_type = params.get("content_type")
+  //   var page_type = content_type == "exhibits.exhibitpage" ? "exhibits" : "collections"
+  //   const resp = await getPageBySlug(page_type, slug)
+  //   const data = await resp.json()
+
+  //   console.log( 'herse hte data though', data )
+  //   if(data && data.id && token && content_type){
+  //     console.log( 'using these values', data.id, content_type, token )
+  //     return await getPreview(data.id, content_type, token)
+  //   }
+  // }
+
+  // leaving existing lookup by id, since current preview link provides only it
   if (params.has('content_type') && params.has('token')) {
     var id = params.get('token').split(':')[0].replace(/id=/g, '')
     return await getPreview(id, params.get('content_type'), params.get('token'))
@@ -18,7 +36,7 @@ export const loader = async ({ request }: LoaderFunctionArgs) => {
 export default function Preview() {
   const preview = useLoaderData()
 
-  if (!preview) {
+  if (!(preview && preview.meta)) {
     throw new Error('Not found!')
   }
